requirements.txt
affine==2.4.0
attrs==25.1.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.1
cycler==0.12.1
efficientnet-pytorch==0.7.1
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Session==0.8.0
fonttools==4.55.8
fsspec==2024.12.0
huggingface-hub==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
msgspec==0.19.0
munch==4.0.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.1.0
pretrainedmodels==0.7.4
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rasterio==1.4.3
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
segmentation_models_pytorch==0.4.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
timm==1.0.14
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
